requirements.txt
pygame
pyopengl
PyOpenGL_accelerate
objloader